Reorganise table constraints into a family directory
Move Table and NegativeTable into gcs/constraints/table/, mirroring how
gcs/constraints/all_different/ is structured. The umbrella table.hh
becomes a thin include-only header; each constraint gets its own
.hh/.cc/_test.cc trio. No algorithmic or behavioural changes.
